Move pipeline code to separate file
